Move menus, add UI for setting auto theme
